feat: implementend new config system
