+++ .env-example
@@ -1,5 +1,5 @@
 REDIS_URL='redis://localhost:6379'
 ENVIRONMENT='development'
-MODE='dynamic'
+MODE='static'
 RUNTIME_TOKEN="<token>"
 FLOW_FALLBACK_PATH='../flow/test_flow_one.json'
