+++ .gitignore
@@ -45,3 +45,5 @@ bin/
 .idea
 
 target
+
+.env
